:truck: Set up ShowPostPage
#195
